requirements.txt
prometheus_client==0.9.0
python-nmap==0.6.1
azure-identity==1.6.1
azure-mgmt-subscription==1.1.0
